backend/requirements.txt
requests
feedparser
psycopg2-binary
google-genai
